+++ src/app.d.ts
@@ -1,7 +1,9 @@
 /// <reference types="bootstrap" />
+/// <reference types="google.accounts" />
 
-// See https://kit.svelte.dev/docs/typescript
+// See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
+// and what to do when importing types
 declare namespace App {
 	interface Locals {
     user: User
@@ -9,11 +11,9 @@ declare namespace App {
 
 	// interface Platform {}
 
-	interface Session {
-    user?: User
-  }
+  // interface PrivateEnv {} // $env/dynamic/private
 
-	// interface Stuff {}
+	// interface PublicEnv {} // $env/dynamic/public
 }
 
 type AuthenticationResult = {