+++ src/routes/auth/reset/[token]/+page.ts
@@ -1,0 +1,7 @@
+import type { PageLoad } from './$types'
+
+export const load: PageLoad = async event => {
+  return {
+    token: event.params.token
+  }
+}